+++ sjsonnet/src/sjsonnet/BaseRenderer.scala
@@ -97,7 +97,7 @@ class BaseRenderer[T <: java.io.Writer]
       case Double.NegativeInfinity => visitString("-Infinity", -1)
       case d if java.lang.Double.isNaN(d) => visitString("NaN", -1)
       case d =>
-        val i = d.toInt
+        val i = d.toLong
         if (d == i) visitFloat64StringParts(i.toString, -1, -1, index)
         else super.visitFloat64(d, index)
         flushBuffer()
